requirements.txt
flask==2.0.3
werkzeug==2.0.3
flask-cors==3.0.10
pyodbc==4.0.32
requests==2.26.0
numpy==1.22.4
scikit-learn==1.1.3
scipy==1.9.3
pandas==1.3.5
python-dotenv==0.19.0
bcrypt==4.0.1
pyjwt==2.3.0
joblib==1.1.0
flask-jwt-extended==4.4.4